Close response.Body when done with the request
Callers should close resp.Body when done reading from it. If resp.Body
is not closed, the Client's underlying RoundTripper (typically
Transport) may not be able to re-use a persistent TCP connection to the
server for a subsequent "keep-alive" request.

Not closing it leads to leaking ile descriptors. Eventually resulting in
exhaustion causing further Net requests to fail.