make go-chef HTTP headers compliant with RFC 2616 (HTTP/1.1) and go1.6
[RFC 2616 § 4.2](https://www.ietf.org/rfc/rfc2616.txt) specifies that HTTP headers **cannot** contain spaces. A space being included in the header is in violation of the HTTP/1.1 standard. Unfortunately, Go 1.6 introduced a change where they outright reject requests, before the your HTTP handler is invoked, if **any** header contains a space in the name:

* https://github.com/golang/go/commit/c052222b3467ca031b267f25f3dcc4b871485793

This changeset alters the `SignRequest` function so that the `Hashed Path` header is no longer sent to the Chef Server. The Chef authentication documentation clearly indicates that this header is **NOT** required to be sent over HTTP as part of authentication. The value is, however, used as part of the generation of the `X-Ops-Authorization` header.