Use type int for Config Timeout
The library expects the given `Timeout` attribute to be set in seconds
and will multiply the value with `time.Second`, but uses the type
time.Duration. The `time.Duration` value is treated as nanoseconds
though. If the library expects the value in seconds, it should use an
appropriate type for that.