Add missing pagination logic in PartialExec
The library has a bug in the PartialExec function where it will only
return the first page of search results.

To address this, I've implemented similar logic to the Exec function to
traverse all pages in the result set and accumulate their rows before
returning. There is enough of a similarity that it may be an indicator
for a need to eventually combine the logic of the two functions.