Update ACL service and documentation
Current version of the Chef server splits actors into two separate
sections for users and clients. The PUT endpoint expects the same
contents so we need to be able to read them.

This API is not included in the Chef documentation.

Code change is based on the knife source code and packet capture.